added a malloc(0) to test-no-malloc
Because of differences in when malloc is called in different architectures, this malloc call is meant to explicitly initialize the allocator.
This will result in an intial allocation of some KB, and brk will be shifted.
This is done to get a consistent reading of brk init across all architectures.
NOTE that malloc(0) does nothing except when allocator is not initialized.
In such a case, it'll call the initializer.